Add start time and time elapsed at start in CSVReader
- timeElapsedAtStart: public variable to define for how long
the data would have been measured before the generator
starts broadcasting the data.

- startTime: private variable to define when the csv reader
started to compare the data time with the time since the
generator started (was enabled) and with the time the whole
application started (in case the reader is not enabled at the
same time the application starts).